iceberg: separate transforms from partitioning
Transforms were previously in partition.cc because they get serialized
as a part of the partition spec. But longer term there will be more
sophisticated logic that is will be needed for transforms specifically
(e.g. applying transforms) so this pulls them out into their own
library.

Also adds day_transform to the transform variant (was defined previously
but not added to the variant).
