+++ src/v/iceberg/transform.cc
@@ -6,7 +6,7 @@
 // As of the Change Date specified in that file, in accordance with
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
-#include "iceberg/partition.h"
+#include "iceberg/transform.h"
 
 namespace iceberg {
